Refactor/Cleanup: move medication request convertion v3-v4 to corresponding file
